requirements.txt
Django~=2.1
djangorestframework~=3.11.0
PyMySQL~=0.9.3